server/requirements.txt
flask
flask-socketio
python-dotenv
gevent
gevent-websocket
llama-index
langchain
